backend/requirements/production.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.9.0
Authlib==1.6.1
billiard==4.2.1
boto3==1.39.16
botocore==1.39.16
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==4.1  
django-parler==2.3
django-parler-rest==2.2
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
kombu==5.5.4
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
Werkzeug==3.1.3
xmltodict==0.14.2